+++ Fighting Game Engine/ScreenManager.h
@@ -34,7 +34,7 @@ public:
 	GLFWwindow* window() const { return _window; }
 
 	const glm::ivec2& screenSize() const { return _screenSize; }
-	const glm::ivec2& screenInvSize() const { return _screenInvSize; }
+	const glm::vec2& screenInvSize() const { return _screenInvSize; }
 
 	const glm::ivec4& viewportSize() const { return _viewportSize; }
